+++ backend/src/main/java/com/aqua/algorithmsDataStructures/services/InterpolationSearchService.java
@@ -27,12 +27,12 @@ public class InterpolationSearchService {
         long memoryAfter = runtime.totalMemory() - runtime.freeMemory();
 
         // calculated time and memory
-        double timeTaken = (double)(endTime - startTime)  / 1000000.0; // divide with 1 000 000 for ms
+        int timeTaken = (int)(endTime - startTime)  / 1000000; // divide with 1 000 000 for ms
         double memoryTaken = (double) (memoryAfter - memoryBefore) / (1024.0 * 1024.0); // divide with 1024*1024 for MB
 
         AlgorithmMetricsResponse response = new AlgorithmMetricsResponse();
         response.setValueFromSearch(valueFromSearch);
-        response.setTimeTaken(timeTaken); // (in ms if divided by 1 000 000)
+        response.setTimeTaken(timeTaken); // (in s if divided by 1000)
         response.setMemoryUsage(memoryTaken);
         return response;
     }
