+++ backend/src/main/java/com/aqua/algorithmsDataStructures/services/LinearSearchService.java
@@ -29,7 +29,7 @@ public AlgorithmMetricsResponse linearSearchMetrics(int[] array, int value) {
     long memoryAfter = runtime.totalMemory() - runtime.freeMemory();
 
     // calculated time and memory
-    double timeTaken = (double)(endTime - startTime)  / 1000000.0; // divide with 1 000 000 for ms
+    int timeTaken = (int)(endTime - startTime)  / 1000000; // divide with 1 000 000 for ms
     double memoryTaken = (double) (memoryAfter - memoryBefore) / (1024.0 * 1024.0); // divide with 1024*1024 for MB
 
     AlgorithmMetricsResponse response = new AlgorithmMetricsResponse();
